checker: extract the rule model into checker/rules
The metadata model (Level, Direction/Area/Kind, Effect, Guards, location
claims, and the Rule struct with its Actions() derivation) moves to a
dedicated package that imports only checker/metaschema. The checker
package binds the model to the check implementations
(BackwardCompatibilityRule embeds rules.Rule and adds Handler) and keeps
the registry, since the handler functions live there.

Type and constant aliases keep the checker package's public surface
unchanged, so external consumers compile as before. ColorMode and the
piped-output detection move along with Level, whose rendering they
serve; the message-argument helpers stay in checker with the change
types that use them.

The spec layer (metaschema + rules) is now importable without the
engine, which is what the audit tests conceptually consume and what
future tooling (the conformance generator, docs generation) needs.
