+++ checker/rules.go
@@ -3,134 +3,70 @@ package checker
 import (
 	"fmt"
 
+	"github.com/oasdiff/oasdiff/checker/rules"
 	"github.com/oasdiff/oasdiff/utils"
 )
 
-type Direction int8
-
-const (
-	DirectionRequest Direction = iota
-	DirectionResponse
-	DirectionNone
+// The rule model (taxonomy, Effect, Guards, Level, location claims) is
+// defined in checker/rules; the aliases below keep the checker package's
+// public surface unchanged. This package binds the model to the check
+// implementations via Handler.
+type (
+	Direction = rules.Direction
+	Area      = rules.Area
+	Kind      = rules.Kind
+	Effect    = rules.Effect
+	Guard     = rules.Guard
 )
 
-// Area is the OpenAPI Object a rule concerns, aligned with the OpenAPI spec's object model.
-type Area int8
-
 const (
-	AreaSchema Area = iota
-	AreaParameters
-	AreaRequestBody
-	AreaResponses
-	AreaPaths
-	AreaHeaders
-	AreaSecurity
-	AreaTags
-	AreaComponents
-	// AreaInfo's rules judge the versioning policy, not the wire contract:
-	// info is annotation-only, so nothing in it can break a client, but
-	// info.version is the declared shape of the change and can contradict it.
-	AreaInfo
-	// AreaServers has no rules yet: server changes are treated as deployment
-	// metadata. It exists so the area taxonomy covers every top-level
-	// document section a future rule could target.
-	AreaServers
-	AreaNone
+	DirectionRequest  = rules.DirectionRequest
+	DirectionResponse = rules.DirectionResponse
+	DirectionNone     = rules.DirectionNone
+
+	AreaSchema      = rules.AreaSchema
+	AreaParameters  = rules.AreaParameters
+	AreaRequestBody = rules.AreaRequestBody
+	AreaResponses   = rules.AreaResponses
+	AreaPaths       = rules.AreaPaths
+	AreaHeaders     = rules.AreaHeaders
+	AreaSecurity    = rules.AreaSecurity
+	AreaTags        = rules.AreaTags
+	AreaComponents  = rules.AreaComponents
+	AreaInfo        = rules.AreaInfo
+	AreaServers     = rules.AreaServers
+	AreaNone        = rules.AreaNone
+
+	KindExistence    = rules.KindExistence
+	KindRequiredness = rules.KindRequiredness
+	KindMutability   = rules.KindMutability
+	KindType         = rules.KindType
+	KindConstraints  = rules.KindConstraints
+	KindValues       = rules.KindValues
+	KindStructure    = rules.KindStructure
+	KindLifecycle    = rules.KindLifecycle
+	KindNone         = rules.KindNone
+
+	EffectNone         = rules.EffectNone
+	EffectWidens       = rules.EffectWidens
+	EffectNarrows      = rules.EffectNarrows
+	EffectIncomparable = rules.EffectIncomparable
+	EffectUnknown      = rules.EffectUnknown
+	EffectViolation    = rules.EffectViolation
+
+	GuardReadOnly   = rules.GuardReadOnly
+	GuardWriteOnly  = rules.GuardWriteOnly
+	GuardSanctioned = rules.GuardSanctioned
+	GuardNonSuccess = rules.GuardNonSuccess
+	GuardHasDefault = rules.GuardHasDefault
+	GuardNegotiated = rules.GuardNegotiated
 )
 
-// Kind is the aspect of the API contract a rule concerns, orthogonal to Area.
-type Kind int8
-
-const (
-	KindExistence Kind = iota
-	KindRequiredness
-	KindMutability
-	KindType
-	KindConstraints
-	KindValues
-	KindStructure
-	KindLifecycle
-	KindNone
-)
-
-func (d Direction) String() string {
-	switch d {
-	case DirectionRequest:
-		return "request"
-	case DirectionResponse:
-		return "response"
-	default:
-		return "none"
-	}
-}
-
-func (a Area) String() string {
-	switch a {
-	case AreaSchema:
-		return "schema"
-	case AreaParameters:
-		return "parameters"
-	case AreaRequestBody:
-		return "requestBody"
-	case AreaResponses:
-		return "responses"
-	case AreaPaths:
-		return "paths"
-	case AreaHeaders:
-		return "headers"
-	case AreaSecurity:
-		return "security"
-	case AreaTags:
-		return "tags"
-	case AreaComponents:
-		return "components"
-	case AreaInfo:
-		return "info"
-	case AreaServers:
-		return "servers"
-	default:
-		return "none"
-	}
-}
-
-func (k Kind) String() string {
-	switch k {
-	case KindExistence:
-		return "existence"
-	case KindRequiredness:
-		return "requiredness"
-	case KindMutability:
-		return "mutability"
-	case KindType:
-		return "type"
-	case KindConstraints:
-		return "constraints"
-	case KindValues:
-		return "values"
-	case KindStructure:
-		return "structure"
-	case KindLifecycle:
-		return "lifecycle"
-	default:
-		return "none"
-	}
-}
-
+// BackwardCompatibilityRule binds the rule metadata to the check function
+// that implements it.
 type BackwardCompatibilityRule struct {
-	Id          string
-	Level       Level
-	Description string
-	Handler     BackwardCompatibilityCheck
-	Direction   Direction
-	Area        Area
-	Kind        Kind
-	Effect      Effect
-	Guards      []Guard
-	// Locations are the metaschema cells this rule covers, as
-	// "location:action[,action...]" claims (see metaschema.ParseClaim).
-	// Empty means the rule is not yet mapped; rule_coverage_test.go reports
-	// unmapped rules and uncovered cells.
-	Locations []string
+	rules.Rule
+	Handler BackwardCompatibilityCheck
 }
 
 func newBackwardCompatibilityRule(id string, level Level, handler BackwardCompatibilityCheck,
@@ -141,16 +77,18 @@ func newBackwardCompatibilityRule(id string, level Level, handler BackwardCompat
 	guards []Guard,
 	locations ...string) BackwardCompatibilityRule {
 	return BackwardCompatibilityRule{
-		Id:          id,
-		Level:       level,
-		Description: descriptionId(id),
-		Handler:     handler,
-		Direction:   direction,
-		Area:        area,
-		Kind:        kind,
-		Effect:      effect,
-		Guards:      guards,
-		Locations:   locations,
+		Rule: rules.Rule{
+			Id:          id,
+			Level:       level,
+			Description: descriptionId(id),
+			Direction:   direction,
+			Area:        area,
+			Kind:        kind,
+			Effect:      effect,
+			Guards:      guards,
+			Locations:   locations,
+		},
+		Handler: handler,
 	}
 }
